Product/requirements.txt
psycopg2
django
djangorestframework
gunicorn